requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
gspread==6.1.2
google-auth==2.30.0
